requirements.txt
# Core dependencies
requests==2.28.1
python-dotenv==0.21.0

# Ollama integration
ollama==0.1.0

# Web scraping and search
beautifulsoup4==4.11.1
searx-api==0.1.0
google-search-results==2.4.1
duckduckgo-search==2.8.5

# Parallel processing
concurrent-log-handler==0.9.20
multiprocessing-logging==0.3.1

# PDF generation
reportlab==3.6.12
fpdf2==2.5.7

# Text processing
nltk==3.7
spacy==3.4.3

# CLI interface
typer==0.7.0
rich==12.6.0
tqdm==4.64.1

# Web interface
flask==2.2.3
werkzeug==2.2.3
jinja2==3.1.2
itsdangerous==2.1.2
markupsafe==2.1.2